% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{dplyr_bundle}
\alias{dplyr_bundle}
\title{A bundle of dplyr functions}
\usage{
dplyr_bundle(input, group, summary)
}
\arguments{
\item{input}{A dataframe or tibble input.}

\item{group}{A categorical column in the dataframe/tibble that should be used for grouping.}

\item{summary}{A numerical column for which summary statistics are to be computed.}
}
\value{
A Tibble containing four columns- categorical variable by which the data is grouped, mean, minimum, maximum computed for a numerical variable.
}
\description{
This function calculates the summary statistics such as mean and range for a numerical variable grouped by a categorical variable (both of which are your choice) from an input dataset.
It uses the dplyr functions group_by and summarise to output a tibble or data frame depending on your input data type class.
}
\examples{
dplyr_bundle(datateachr::vancouver_trees, group = common_name, summary = diameter)
dplyr_bundle(palmerpenguins::penguins, group = island, summary = flipper_length_mm)
}
